task7

+++ dao/src/main/java/com/netcracker/denisik/entities/Student.java
@@ -17,7 +17,7 @@ public class Student extends User{
     private long id;
 
     @Column(name = "groupId")
-    private int groupId;
+    private long groupId;
 
     @ManyToOne
     @JoinColumn(name = "specialities_id")
@@ -28,7 +28,7 @@ public class Student extends User{
     private WriteBook writeBook;
 
     @Builder(builderMethodName = "builderStudent")
-    public Student(User user, long id, int groupId, Speciality speciality, WriteBook writeBook) {
+    public Student(User user, long id, long groupId, Speciality speciality, WriteBook writeBook) {
         super(id, user.getRole(), user.getPassword(), user.getLogin(), user.getName());
         this.id = id;
         this.groupId = groupId;
